feat: :art: se mejora la estructura y organizaicon de los componentes
se agrega el boton para mostrar los mensajes en la consulta general
