+++ src/components/DashboardPage/DashboardTable.jsx
@@ -1,12 +1,12 @@
 import React, { useState } from "react";
 import { FiltrosReportes } from './FiltrosReportes';
 import { TarjetaContador } from './TarjetaContador';
+import { UserMessagesModal } from './UserMessagesModal';
 
 export const DashboardTable = () => {
-
-  const [selectedUser, setSelectedUser] = useState(null); // Usuario seleccionado para el modal
-  const [isModalOpen, setIsModalOpen] = useState(false); // Estado del modal
-  const [filteredData, setFilteredData] = useState([]); // Usar los datos filtrados o el JSON simulado
+  const [selectedUser, setSelectedUser] = useState(null);
+  const [isModalOpen, setIsModalOpen] = useState(false);
+  const [filteredData, setFilteredData] = useState([]);
 
   const openModal = (user) => {
     setSelectedUser(user);
@@ -18,9 +18,8 @@ export const DashboardTable = () => {
     setIsModalOpen(false);
   };
 
-  // Función que maneja los datos obtenidos después de aplicar los filtros
   const handleDataFetched = (fetchedData) => {
-    console.log(fetchedData);
+    console.log('informacion de isadata ',fetchedData);
     setFilteredData(fetchedData);
   };
 
@@ -88,53 +87,12 @@ export const DashboardTable = () => {
         </table>
       </div>
 
-      {/* Modal */}
-      {isModalOpen && selectedUser && (
-        <div
-          className="fixed inset-0 flex items-center justify-center bg-black/50 z-50"
-          onClick={(e) => {
-            // Verificar si el clic ocurrió fuera del modal
-            if (e.target === e.currentTarget) {
-              closeModal();
-            }
-          }}
-        >
-          <div className="bg-[#162033] p-6 rounded-lg shadow-lg text-white h-[420px] w-[920px] overflow-y-scroll scrollbar-custom">
-            {/* Contenedor para el título y el botón */}
-            <div className="flex flex-row justify-between items-center mb-4">
-              <h3 className="text-xl font-bold">Mensajes de {selectedUser.Username}</h3>
-              <button
-                className="px-4 py-2 bg-red-400 text-white rounded-lg hover:bg-red-500 transition-all duration-300"
-                onClick={closeModal}
-              >
-                Cerrar
-              </button>
-            </div>
-
-            {/* Tabla de mensajes */}
-            <table className="min-w-full text-left">
-              <thead>
-                <tr className="border-b border-cyan-400">
-                  <th className="px-4 py-2">Message</th>
-                  <th className="px-4 py-2">ID</th>
-                  <th className="px-4 py-2">Time</th>
-                </tr>
-              </thead>
-              <tbody>
-                {selectedUser.Messages.map((message) => (
-                  <tr key={message.MessageId} className="border-b border-gray-600 hover:bg-[#3B3F47]">
-                    <td className="px-4 py-2">{message.Message}</td>
-                    <td className="px-4 py-2">{message.MessageId}</td>
-                    <td className="px-4 py-2">{message.Time}</td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          </div>
-        </div>
-      )}
-
-
+      {/* Modal Component */}
+      <UserMessagesModal 
+        isOpen={isModalOpen}
+        user={selectedUser}
+        onClose={closeModal}
+      />
     </div>
   );
 };
